Python/StorageAccountBlobHelper/requirements.txt
azure-identity>=1.13.0
azure-storage-blob>=12.19.0
